mercurial: checkout target commit before get_diff (Bug 1926924) r=sheehan
This step was removed from the main submit logic, as it was no longer
needed by git, and looked unecessary for hg. Not so!

When computing the diffs, the Hg logic seems to take into account the
current state of the working directory. This leads to incorrect diffs
for anything but the last commit of a stack.

Differential Revision: https://phabricator.services.mozilla.com/D226848